pts边界值校验

+++ services/media_engine/plugins/ffmpeg_adapter/demuxer/block_queue_pool.h
@@ -33,7 +33,7 @@ namespace OHOS {
 namespace Media {
 
 struct SamplePacket {
-    uint64_t offset = 0;
+    uint32_t offset = 0;
     AVPacket* pkt = nullptr;
     bool isEOS = false;
     ~SamplePacket()
